VRProject
>A project based on Virtual Reality.
>We are developing a game which will be able to run on any Head Mounted
device based on the concept of Google Cardboard.
>We are using Unity3D 4.6.1 to develop our virtually real environment.
>As per the project requirements, we need to to add networking
capabilities to the game. Multiple people present in the same VR
environment should be able to interact in any of the possible ways.
>For networking, Instead of writing native socket code and to prevent us
from the tedious job of network backend coding(which is not our motive)
, we will be using "Nextpeer SDK" for Unity which provide enhanced
social and networking features for the game.